Adding transaction's service and controller

+++ src/main/java/com/project/bank/model/Transaction.java
@@ -5,6 +5,7 @@ import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 import java.time.LocalDateTime;
 
@@ -21,7 +22,7 @@ public class Transaction {
     @ManyToOne
     @JoinColumn(name = "account_id")
     private Account account;
-    private float amount;
+    private BigDecimal amount;
     private String type;
     private LocalDateTime date;
 }
